+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Add safeguard to avoid referencing undefined variable
+
 ## [1.2.0] - 2019-07-17
 
 ### Added